🎈 perf: 优化运行逻辑

+++ bin/run.sh
@@ -1,2 +1,15 @@
-rm -r py_cpp_pipe.fifo
-./monitor -i address.conf & python write.py
+if [ -d "logs" ]; then
+    echo "logs directory already exists."
+    rm -r logs
+else
+    echo "Created logs directory."
+fi
+
+mkdir logs
+
+if [ -d "py_cpp_pipe.fifo" ]; then
+    echo "Build directory already exists."
+    rm -r py_cpp_pipe.fifo
+fi
+
+python3 write.py & ./monitor -i address.conf
